Lesson 5: split to seperate modules
Lesson 5. HTML parcing. Scrapy

Move the queries dictionary,
which the spider autoyoula uses in parse methods and item template,
to a separate parse.lib.autoyoula module.

Move the decoder function
for extracting of user id from the script,
which the autoyoula spider uses in the item template,
to a separate parse.lib.funcs module.

Add a gitignore file
with special exeception to track /parse/lib path.

Changes to be committed:
    new file:   lesson_4/.gitignore
    modified:   lesson_4/main.py
    new file:   lesson_4/parse/lib/__init__.py
    new file:   lesson_4/parse/lib/autoyoula.py
    new file:   lesson_4/parse/lib/funcs.py
    modified:   lesson_4/parse/pipelines.py
    modified:   lesson_4/parse/spiders/autoyoula.py